testing: Fix console_log tests
Bring in the pycloudlib change effecting the console_log on ec2.
Additionally, provide a helper function for retrieving the console log
to reduce duplication.

Moved the retry decorator into a new file to avoid cyclic dependency
and had to update call sites accordingly.
