+++ integration-requirements.txt
@@ -1,5 +1,5 @@
 # PyPI requirements for cloud-init integration testing
 # https://cloudinit.readthedocs.io/en/latest/topics/integration_tests.html
 #
-pycloudlib @ git+https://github.com/canonical/pycloudlib.git@db36ef2dfc0a8d916c0e2c1794d7228d49de9192
+pycloudlib @ git+https://github.com/canonical/pycloudlib.git@a8584c9da7ef7648cd8702d96a5696de6a6de0e3
 pytest
