+++ tests/integration_tests/modules/test_keys_to_console.py
@@ -4,7 +4,9 @@
 ``tests/cloud_tests/testcases/modules/keys_to_console.yaml``.)"""
 import pytest
 
-from tests.integration_tests.util import retry
+from tests.integration_tests.decorators import retry
+from tests.integration_tests.instances import IntegrationInstance
+from tests.integration_tests.util import get_console_log
 
 BLACKLIST_USER_DATA = """\
 #cloud-config
@@ -85,29 +87,10 @@ class TestKeysToConsoleDisabled:
 
 
 @pytest.mark.user_data(ENABLE_KEYS_TO_CONSOLE_USER_DATA)
-@pytest.mark.ec2
-@pytest.mark.lxd_container
-@pytest.mark.oci
-@pytest.mark.openstack
-class TestKeysToConsoleEnabled:
+@retry(tries=30, delay=1)
+def test_duplicate_messaging_console_log(client: IntegrationInstance):
     """Test that output can be enabled disabled."""
-
-    def test_duplicate_messaging_console_log(self, class_client):
-        class_client.execute("cloud-init status --wait --long").ok
-        try:
-            console_log = class_client.instance.console_log()
-        except NotImplementedError:
-            # Assume that an exception here means that we can't use the console
-            # log
-            pytest.skip("NotImplementedError when requesting console log")
-            return
-        if console_log.lower() == "no console output":
-            # This test retries because we might not have the full console log
-            # on the first fetch. However, if we have no console output
-            # at all, we don't want to keep retrying as that would trigger
-            # another 5 minute wait on the pycloudlib side, which could
-            # leave us waiting for a couple hours
-            pytest.fail("no console output")
-            return
-        msg = "no authorized SSH keys fingerprints found for user barfoo."
-        assert 1 == console_log.count(msg)
+    assert (
+        "no authorized SSH keys fingerprints found for user barfoo."
+        in get_console_log(client)
+    )
