+++ tests/integration_tests/util.py
@@ -1,4 +1,3 @@
-import functools
 import logging
 import multiprocessing
 import os
@@ -7,6 +6,10 @@ import time
 from collections import namedtuple
 from contextlib import contextmanager
 from pathlib import Path
+
+import pytest
+
+from tests.integration_tests.instances import IntegrationInstance
 
 log = logging.getLogger("integration_testing")
 key_pair = namedtuple("key_pair", "public_key private_key")
@@ -115,33 +118,11 @@ def get_test_rsa_keypair(key_name: str = "test1") -> key_pair:
     return key_pair(public_key, private_key)
 
 
-def retry(*, tries: int = 30, delay: int = 1):
-    """Decorator for retries.
-
-    Retry a function until code no longer raises an exception or
-    max tries is reached.
-
-    Example:
-      @retry(tries=5, delay=1)
-      def try_something_that_may_not_be_ready():
-          ...
-    """
-
-    def _retry(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            last_error = None
-            for _ in range(tries):
-                try:
-                    func(*args, **kwargs)
-                    break
-                except Exception as e:
-                    last_error = e
-                    time.sleep(delay)
-            else:
-                if last_error:
-                    raise last_error
-
-        return wrapper
-
-    return _retry
+def get_console_log(client: IntegrationInstance):
+    try:
+        console_log = client.instance.console_log()
+    except NotImplementedError:
+        pytest.skip("NotImplementedError when requesting console log")
+    if console_log.lower().startswith("no console output"):
+        pytest.fail("no console output")
+    return console_log
